This commit mainly adds styling for individual task pages.
Individual Task Pages have been changed from a top to bottom view to a left to right view.
Custom checkboxes have been implemented. In order to use them, the following syntax should be used:
```
<label for = "_" class="form_control">
    <input type="checkbox">
    <!-- Label Text can go here -->
</label>
```

It also changes the structure of tasks that are received from DayMap and Google Classroom.
Instead of the grade being returned as a string, it is now returned as a struct named "TaskGrade"

This is represented in the individual task page as a letter grade, a mark percentage, and a color based on that percentage.
The values will default to `"-"`, `0.0`, and `#ffffff` respectively.

A new package that holds the key structs that are used in the daymap and google classroom packages, as well as the main package has been added in order to fix some issues with struct conversion.
This package has been named "plat", short for "platform", and the structs have been placed in `plat/types.go`.