feat: Add TaskCollect logos and favicons
TaskCollect now has an official logo that will be used on the website
and other TaskCollect-related endeavours. A normal version of the logo
and wordmark are what will be used primarily. However, there exists a
"fancy" version of both the logo and wordmark that can be used for
promotional purposes.

New:
- TaskCollect logo on the login screen and navbar.
- Favicon