Improve timetable experience on mobile
On mobile, all days before the current day are not displayed.
This is to ensure the day at the top of the screen is the current day, for easier navigation.
The top margin for each lesson block has been increased to the size of the bottom margin for aesthetic appeal.