Display timetable correctly in mobile view on weekends
In mobile view, only the days after and including the current day are displayed in the timetable.
However, due to an issue with TaskCollect's logic, this meant that no day of the upcoming week was displayed in the timetable during weekends (because the days in the weekend follow weekdays).
This has now been corrected and during weekends, all days of the upcoming week are displayed in mobile view.