Render posted dates separately to due dates
Posted dates are read with consideration of the past ("posted Tuesday" is equivalent to "posted last Tuesday").
Due dates are read with consideration of the future ("due Tuesday" is equivalent to "due this upcoming Tuesday").
Therefore it is necessary to render posted dates and due dates separately, a differentiation which is introduced in this commit.